:heavy_plus_sign: Add Metadata class
