+++ .pylintrc
@@ -14,4 +14,6 @@ ignore=docs
 # --disable=W"
 
 # we're going to support as many arguments as the API calls have
-disable=too-many-arguments
+disable=too-many-arguments,
+        # this is be removed after refactoring
+        duplicate-code
